requirements.txt
yfinance
numpy
dotenv
requests
pydantic
finnhub-python


